Rename `music-constants.ts` to `music.ts`
